Add docker client to 'thick' image
The `gcloud docker` subcommand relies on executing a docker client.
This change allows the user to bindmount in the host's docker socket,
and then interact with it using the `gcloud docker` subcommand.

This will only work if the user's host dockerd is compatible with this
docker client, but this is true for all modern versions of docker.

See additional discussion in https://github.com/GoogleCloudPlatform/cloud-sdk-docker/issues/97
for why bindmounting in the client's docker binary is non-ideal.

The `docker` image was used to source the docker client binary as a
matter of convenience.

The change is intentionally only applied to the 'debian' variant of the
image, not the 'alpine' nor '-slim' ones. This is to keep the 'slim'
image slim.