Add an 'it runs' failsafes to binaries
This adds a "the binary runs" check to the end of each container's list
of commands.

The primary benefit this has is that a build will fail if any of those
commands don't work, which is preferable to a bad release being made.

As a small secondary benefit, it also includes the `version` output in
the build log.

Ideally, a more complete functional test would be done on each image
prior to push, but since dockerhub automated builds are being used,
there's no great place to insert those tests.

I think this is enough of an improvement to be worth doing.