Update Dockerfile for `:latest` and `:debian_component_based` images to remove `make` package (#598)
* Update Dockerfile for `:latest` and `:debian_component_based` images to remove `make` package

We are removing the `make` dependency from the `:latest` and `:debian_component_based` Google Cloud CLI Docker images to mitigate customers' exposure to vulnerabilities found in this component and its dependencies according to the [following timeline](#update-timeline). If your workflows rely on the `make` package, you will need to pin to the respective `Pin-To` gcloud version or earlier. Alternatively, you could build your own docker image and include `make` using a custom Dockerfile. 

### Update Timeline

|  <img width=100/> Date  | Removed in gcloud version | `Pin-to` gcloud version to continue using `make` | `make` removed from images |
|:----------:|:-------------------------------------------:|:--------------------:|:----------:|
| Aug 5, 2025 | 533.0.0 | 532.0.0 | `:latest` and `:debian_component_based` |

* Update Dockerfile